Refactor maps construction in AoC2023 Day 5 puzzle
This commit refactors the method 'construct_maps' in the file 'if_you_give_a_seed_a_fertilizer.rb' of Advent of Code 2023 Day 5 puzzle. It simplifies the loop by rejecting empty lines before the loop starts, instead of having a conditional inside the loop to deal with them. This makes the code more readable and efficient.