Implement MirageMaintenance puzzle solution and tests.
This commit introduces the implementation of the MirageMaintenance class representing a puzzle solution. The class is capable of predicting sequence terms and providing answers for both part one and part two of the puzzle. In addition, a corresponding test suite has been created to verify the accuracy of the solution with both provided and actual input data.