Add StepCounter class and corresponding tests
This commit introduces a new class `StepCounter` for AoC2023 puzzles. It also includes the tests for `StepCounter` class in `step_counter_spec.rb`. The main functionality of this class is to read from an input file and calculate the count of steps made by an elf across a garden plot within a given number of steps.